[symfony-insight] Refactor categories: Symfony controller action method should not be too long, Source code should not contain TODO comments
